+++ src/thinpi-http/reboot.c
@@ -1,8 +1,0 @@
-#include <stdio.h>
-
-int main() {
-	
-	system("sudo reboot");
-	
-	return 0;
-}