+++ src/thinpi-http/shutdown.c
@@ -1,8 +1,0 @@
-#include <stdio.h>
-
-int main() {
-	
-	system("sudo shutdown -h now");
-	
-	return 0;
-}