+++ github/CHANGELOG.md
@@ -15,6 +15,7 @@ All notable changes to this project will be documented in this file.
 - [CONFIG] Added ThinPi Settings Page - **WIP**
 - [CONFIG] Editing Servers in the list from the GUI Config Manager now works (see BUGS.md for more info TPBUG#001)
 - [CONFIG] Adding Servers from the GUI Config Manager now works (see BUGS.md for more info TPBUG#001)
+- [HTTP] Removed thinpi-http code, as with the new updates it is broken.
 
 ## 0.2.0 - RC